requirements.txt
boto3
python-dotenv
pymupdf
langchain
langchain-community
nltk
llama-parse
nest_asyncio
pandas
faiss-cpu